+++ ovos_stt_plugin_server/__init__.py
@@ -1,9 +1,10 @@
-from queue import Queue
-from uuid import uuid4
+from typing import Optional, List
 
 import requests
+import random
+from requests.utils import default_user_agent
+from ovos_utils.log import LOG
 from ovos_plugin_manager.stt import STT
-from ovos_plugin_manager.stt import StreamingSTT, StreamThread
 
 
 class OVOSHTTPServerSTT(STT):
@@ -11,263 +12,170 @@ class OVOSHTTPServerSTT(STT):
 
     def __init__(self, config=None):
         super().__init__(config)
-        self.url = self.config.get("url") or "https://stt.strongthany.cc/stt"
+        if not self.verify_ssl:
+            LOG.warning("SSL verification disabled, this is not secure and should"
+                        "only be used for test systems! Please set up a valid certificate!")
+
+    @property
+    def verify_ssl(self) -> bool:
+        return self.config.get("verify_ssl", True)
+
+    @property
+    def user_agent(self) -> str:
+        return self.config.get("user_agent") or default_user_agent()
+
+    @property
+    def public_servers(self):
+        return [
+            "https://fasterwhisper.ziggyai.online/stt",
+            "https://stt.smartgic.io/fasterwhisper/stt",
+            "https://whisper.neonaiservices.com/stt"
+        ]
+
+    @property
+    def urls(self) -> Optional[List[str]]:
+        urls = self.config.get("url", self.config.get("urls"))
+        if urls and not isinstance(urls, list):
+            urls = [urls]
+        return urls
 
     def execute(self, audio, language=None):
-        self.response = requests.post(self.url, data=audio.get_wav_data(),
-                                      headers={"Content-Type": "audio/wav"},
-                                      params={"lang": language or self.lang})
-        return self.response.text if self.response else None
+        if self.urls:
+            LOG.debug(f"Using user defined urls {self.urls}")
+            urls = self.urls
+        else:
+            LOG.debug(f"Using public servers {self.public_servers}")
+            urls = self.public_servers
+            random.shuffle(urls)
+        for url in urls:
+            LOG.debug(f"chosen url {url}")
+            try:
+                response = requests.post(url, data=audio.get_wav_data(),
+                                         headers={"Content-Type": "audio/wav",
+                                                  "User-Agent": self.user_agent},
+                                         params={"lang": language or self.lang},
+                                         verify=self.verify_ssl)
+                if not response.ok:
+                    LOG.error(f"{response.status_code} response from {url}: "
+                              f"{response.content}")
+                else:
+                    return response.text
+            except Exception as e:
+                LOG.exception(e)
+            LOG.error(f"STT request to {url} failed")
 
 
-class OVOSHTTPStreamServerStreamThread(StreamThread):
-    def __init__(self, queue, language, url="https://stt.strongthany.cc/stream"):
-        super().__init__(queue, language)
-        self.url = url
-        self.session = requests.Session()
-
-    def reset_model(self, session_id=None):
-        self.session_id = session_id or str(uuid4())
-        # reset the model for this session
-        response = self.session.post(f"{self.url}/start",
-                                     params={"lang": self.language,
-                                             "uuid": self.session_id})
-
-    def handle_audio_stream(self, audio, language):
-        lang = language or self.language
-        response = self.session.post(f"{self.url}/audio",
-                                     params={"lang": lang,
-                                             "uuid": self.session_id},
-                                     data=audio, stream=True)
-        self.text = response.json()["transcript"]
-        return self.text
-
-    def finalize(self):
-        """ return final transcription """
-        try:
-            response = self.session.post(f"{self.url}/end",
-                                         params={"lang": self.language,
-                                                 "uuid": self.session_id})
-            self.text = response.json()["transcript"] or self.text
-        except:
-            pass
-        return self.text
-
-
-class OVOSHTTPStreamServerSTT(StreamingSTT):
-    """Streaming STT interface for the OVOS-HTTP-STT-Server"""
-
-    def create_streaming_thread(self):
-        url = self.config.get('url') or "https://stt.strongthany.cc/stream"
-        self.queue = Queue()
-
-        stream = OVOSHTTPStreamServerStreamThread(self.queue, self.lang, url)
-        stream.reset_model()
-        return stream
-
-
-# will list the public instances of google STT proxies as valid configs
-# but thats not the main intended usage of this plugin
-
-
-# taken from https://stackoverflow.com/questions/14257598/what-are-language-codes-in-chromes-implementation-of-the-html5-speech-recogniti/14302134#14302134
-_lang = {
-    "Afrikaans": [
-        ["South Africa", "af-ZA"]
-    ],
-    "Arabic": [
-        ["Algeria", "ar-DZ"],
-        ["Bahrain", "ar-BH"],
-        ["Egypt", "ar-EG"],
-        ["Israel", "ar-IL"],
-        ["Iraq", "ar-IQ"],
-        ["Jordan", "ar-JO"],
-        ["Kuwait", "ar-KW"],
-        ["Lebanon", "ar-LB"],
-        ["Morocco", "ar-MA"],
-        ["Oman", "ar-OM"],
-        ["Palestinian Territory", "ar-PS"],
-        ["Qatar", "ar-QA"],
-        ["Saudi Arabia", "ar-SA"],
-        ["Tunisia", "ar-TN"],
-        ["UAE", "ar-AE"]
-    ],
-    "Basque": [
-        ["Spain", "eu-ES"]
-    ],
-    "Bulgarian": [
-        ["Bulgaria", "bg-BG"]
-    ],
-    "Catalan": [
-        ["Spain", "ca-ES"]
-    ],
-    "Chinese Mandarin": [
-        ["China (Simp.)", "cmn-Hans-CN"],
-        ["Hong Kong SAR (Trad.)", "cmn-Hans-HK"],
-        ["Taiwan (Trad.)", "cmn-Hant-TW"]
-    ],
-    "Chinese Cantonese": [
-        ["Hong Kong", "yue-Hant-HK"]
-    ],
-    "Croatian": [
-        ["Croatia", "hr_HR"]
-    ],
-    "Czech": [
-        ["Czech Republic", "cs-CZ"]
-    ],
-    "Danish": [
-        ["Denmark", "da-DK"]
-    ],
-    "English": [
-        ["Australia", "en-AU"],
-        ["Canada", "en-CA"],
-        ["India", "en-IN"],
-        ["Ireland", "en-IE"],
-        ["New Zealand", "en-NZ"],
-        ["Philippines", "en-PH"],
-        ["South Africa", "en-ZA"],
-        ["United Kingdom", "en-GB"],
-        ["United States", "en-US"]
-    ],
-    "Farsi": [
-        ["Iran", "fa-IR"]
-    ],
-    "French": [
-        ["France", "fr-FR"]
-    ],
-    "Filipino": [
-        ["Philippines", "fil-PH"]
-    ],
-    "Galician": [
-        ["Spain", "gl-ES"]
-    ],
-    "German": [
-        ["Germany", "de-DE"]
-    ],
-    "Greek": [
-        ["Greece", "el-GR"]
-    ],
-    "Finnish": [
-        ["Finland", "fi-FI"]
-    ],
-    "Hebrew": [
-        ["Israel", "he-IL"]
-    ],
-    "Hindi": [
-        ["India", "hi-IN"]
-    ],
-    "Hungarian": [
-        ["Hungary", "hu-HU"]
-    ],
-    "Indonesian": [
-        ["Indonesia", "id-ID"]
-    ],
-    "Icelandic": [
-        ["Iceland", "is-IS"]
-    ],
-    "Italian": [
-        ["Italy", "it-IT"],
-        ["Switzerland", "it-CH"]
-    ],
-    "Japanese": [
-        ["Japan", "ja-JP"]
-    ],
-    "Korean": [
-        ["Korea", "ko-KR"]
-    ],
-    "Lithuanian": [
-        ["Lithuania", "lt-LT"]
-    ],
-    "Malaysian": [
-        ["Malaysia", "ms-MY"]
-    ],
-    "Dutch": [
-        ["Netherlands", "nl-NL"]
-    ],
-    "Norwegian": [
-        ["Norway", "nb-NO"]
-    ],
-    "Polish": [
-        ["Poland", "pl-PL"]
-    ],
-    "Portuguese": [
-        ["Brazil", "pt-BR"],
-        ["Portugal", "pt-PT"]
-    ],
-    "Romanian": [
-        ["Romania", "ro-RO"]
-    ],
-    "Russian": [
-        ["Russia", "ru-RU"]
-    ],
-    "Serbian": [
-        ["Serbia", "sr-RS"]
-    ],
-    "Slovak": [
-        ["Slovakia", "sk-SK"]
-    ],
-    "Slovenian": [
-        ["Slovenia", "sl-SI"]
-    ],
-    "Spanish": [
-        ["Argentina", "es-AR"],
-        ["Bolivia", "es-BO"],
-        ["Chile", "es-CL"],
-        ["Colombia", "es-CO"],
-        ["Costa Rica", "es-CR"],
-        ["Dominican Republic", "es-DO"],
-        ["Ecuador", "es-EC"],
-        ["El Salvador", "es-SV"],
-        ["Guatemala", "es-GT"],
-        ["Honduras", "es-HN"],
-        ["México", "es-MX"],
-        ["Nicaragua", "es-NI"],
-        ["Panamá", "es-PA"],
-        ["Paraguay", "es-PY"],
-        ["Perú", "es-PE"],
-        ["Puerto Rico", "es-PR"],
-        ["Spain", "es-ES"],
-        ["Uruguay", "es-UY"],
-        ["United States", "es-US"],
-        ["Venezuela", "es-VE"]
-    ],
-    "Swedish": [
-        ["Sweden", "sv-SE"]
-    ],
-    "Thai": [
-        ["Thailand", "th-TH"]
-    ],
-    "Turkish": [
-        ["Turkey", "tr-TR"]
-    ],
-    "Ukrainian": [
-        ["Ukraine", "uk-UA"]
-    ],
-    "Vietnamese": [
-        ["Viet Nam", "vi-VN"]
-    ],
-    "Zulu": [
-        ["South Africa", "zu-ZA"]
-    ]
+_whisper_lang = {
+    "en": "english",
+    "zh": "chinese",
+    "de": "german",
+    "es": "spanish",
+    "ru": "russian",
+    "ko": "korean",
+    "fr": "french",
+    "ja": "japanese",
+    "pt": "portuguese",
+    "tr": "turkish",
+    "pl": "polish",
+    "ca": "catalan",
+    "nl": "dutch",
+    "ar": "arabic",
+    "sv": "swedish",
+    "it": "italian",
+    "id": "indonesian",
+    "hi": "hindi",
+    "fi": "finnish",
+    "vi": "vietnamese",
+    "iw": "hebrew",
+    "uk": "ukrainian",
+    "el": "greek",
+    "ms": "malay",
+    "cs": "czech",
+    "ro": "romanian",
+    "da": "danish",
+    "hu": "hungarian",
+    "ta": "tamil",
+    "no": "norwegian",
+    "th": "thai",
+    "ur": "urdu",
+    "hr": "croatian",
+    "bg": "bulgarian",
+    "lt": "lithuanian",
+    "la": "latin",
+    "mi": "maori",
+    "ml": "malayalam",
+    "cy": "welsh",
+    "sk": "slovak",
+    "te": "telugu",
+    "fa": "persian",
+    "lv": "latvian",
+    "bn": "bengali",
+    "sr": "serbian",
+    "az": "azerbaijani",
+    "sl": "slovenian",
+    "kn": "kannada",
+    "et": "estonian",
+    "mk": "macedonian",
+    "br": "breton",
+    "eu": "basque",
+    "is": "icelandic",
+    "hy": "armenian",
+    "ne": "nepali",
+    "mn": "mongolian",
+    "bs": "bosnian",
+    "kk": "kazakh",
+    "sq": "albanian",
+    "sw": "swahili",
+    "gl": "galician",
+    "mr": "marathi",
+    "pa": "punjabi",
+    "si": "sinhala",
+    "km": "khmer",
+    "sn": "shona",
+    "yo": "yoruba",
+    "so": "somali",
+    "af": "afrikaans",
+    "oc": "occitan",
+    "ka": "georgian",
+    "be": "belarusian",
+    "tg": "tajik",
+    "sd": "sindhi",
+    "gu": "gujarati",
+    "am": "amharic",
+    "yi": "yiddish",
+    "lo": "lao",
+    "uz": "uzbek",
+    "fo": "faroese",
+    "ht": "haitian creole",
+    "ps": "pashto",
+    "tk": "turkmen",
+    "nn": "nynorsk",
+    "mt": "maltese",
+    "sa": "sanskrit",
+    "lb": "luxembourgish",
+    "my": "myanmar",
+    "bo": "tibetan",
+    "tl": "tagalog",
+    "mg": "malagasy",
+    "as": "assamese",
+    "tt": "tatar",
+    "haw": "hawaiian",
+    "ln": "lingala",
+    "ha": "hausa",
+    "ba": "bashkir",
+    "jw": "javanese",
+    "su": "sundanese",
 }
 
-OVOSHTTPServerSTTConfig = {}
+if __name__ == "__main__":
+    from speech_recognition import Recognizer, AudioFile
 
-for lang, data in _lang.items():
-    for region, code in data:
-        OVOSHTTPServerSTTConfig[code] = [
-            {"lang": code,
-             "url": "https://stt.openvoiceos.com",
-             "meta": {
-                 "priority": 30,
-                 "display_name": f"OVOS Google Proxy {lang} ({region})",
-                 "offline": False}
-             },
-            {"lang": code,
-             "url": "https://stt.strongthany.cc",
-             "meta": {
-                 "priority": 80,
-                 "display_name": f"Strongthany Google Proxy {lang} ({region})",
-                 "offline": False}
-             }
-        ]
+    engine = OVOSHTTPServerSTT()
+
+    # inference
+    jfk = "/home/miro/PycharmProjects/ovos-stt-plugin-fasterwhisper/jfk.wav"
+    with AudioFile(jfk) as source:
+        audio = Recognizer().record(source)
+
+    pred = engine.execute(audio)
+    print(pred)